moved UI scripts, deleted unused script
