--- Assets/Scripts/UI/Drag.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4f3df0a4944493247b415f6072bb43cc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
