+++ Assets/_Application/Cameras/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1031ae48cd6f543b7b2accaf9020a67a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
